+++ clickhouse-cli-client/src/test/java/com/clickhouse/client/cli/ClickHouseCommandLineClientTest.java
@@ -11,6 +11,7 @@ import com.clickhouse.client.cli.config.ClickHouseCommandLineOption;
 import com.clickhouse.data.ClickHouseCompression;
 
 import java.io.IOException;
+import java.util.concurrent.ExecutionException;
 
 import org.testcontainers.containers.GenericContainer;
 import org.testng.Assert;
@@ -18,7 +19,8 @@ import org.testng.SkipException;
 import org.testng.annotations.BeforeClass;
 import org.testng.annotations.DataProvider;
 import org.testng.annotations.Test;
-
+// deprecate from version 0.6.0
+@Deprecated
 public class ClickHouseCommandLineClientTest extends ClientIntegrationTest {
     @BeforeClass
     static void init() {
@@ -182,5 +184,10 @@ public class ClickHouseCommandLineClientTest extends ClientIntegrationTest {
     @Override
     public void testTransactionTimeout() throws ClickHouseException {
         throw new SkipException("Skip due to session is not supported");
+    }
+    @Test(groups = { "integration" })
+    @Override
+    public void testRowBinaryWithDefaults() throws ClickHouseException, IOException, ExecutionException, InterruptedException {
+        throw new SkipException("Skip due to supported");
     }
 }