+++ clickhouse-grpc-client/src/main/java/com/clickhouse/client/grpc/ClickHouseGrpcClientImpl.java
@@ -47,7 +47,7 @@ import com.clickhouse.data.ClickHousePipedOutputStream;
 import com.clickhouse.data.ClickHouseUtils;
 import com.clickhouse.logging.Logger;
 import com.clickhouse.logging.LoggerFactory;
-
+@Deprecated
 public class ClickHouseGrpcClientImpl extends AbstractClient<ManagedChannel> {
     private static final Logger log = LoggerFactory.getLogger(ClickHouseGrpcClientImpl.class);