+++ clickhouse-grpc-client/src/main/java/com/clickhouse/client/grpc/config/ClickHouseGrpcOption.java
@@ -8,6 +8,7 @@ import com.clickhouse.data.ClickHouseChecker;
 /**
  * gRPC client options.
  */
+@Deprecated
 public enum ClickHouseGrpcOption implements ClickHouseOption {
     /**
      * Flow control window.