+++ clickhouse-http-client/src/main/java/com/clickhouse/client/http/ClickHouseHttpConnection.java
@@ -164,10 +164,9 @@ public abstract class ClickHouseHttpConnection implements AutoCloseable {
     }
 
     static String buildUrl(String baseUrl, ClickHouseRequest<?> request) {
-        ClickHouseConfig config = request.getConfig();
-
         StringBuilder builder = new StringBuilder().append(baseUrl);
-        String context = config.getStrOption(ClickHouseHttpOption.WEB_CONTEXT);
+        // TODO: Using default until we will remove
+        String context = "/";
         if (!ClickHouseChecker.isNullOrEmpty(context)) {
             char prev = '/';
             for (int i = 0, len = context.length(); i < len; i++) {
@@ -366,6 +365,7 @@ public abstract class ClickHouseHttpConnection implements AutoCloseable {
         this.config = c;
         this.defaultHeaders = Collections.unmodifiableMap(createDefaultHeaders(c, server, getUserAgent()));
         this.url = buildUrl(server.getBaseUri(), request);
+        log.debug("url [%s]", this.url);
     }
 
     protected void closeQuietly() {